Use u32 for degree, fix impl Arbitrary for Instance

+++ rust/ommx/src/convert/linear.rs
@@ -52,7 +52,7 @@ impl Linear {
         self.terms.iter().map(|term| term.id).collect()
     }
 
-    pub fn degree(&self) -> usize {
+    pub fn degree(&self) -> u32 {
         if self.terms.is_empty() {
             0
         } else {
